requirements.txt
streamlit
yfinance
pandas
numpy
pyyaml
